tensortrade/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
ib_insync
# Add dotenv for environment variable management
python-dotenv
# Add nest_asyncio for running asyncio loops in scripts/notebooks
nest_asyncio
tensortrade
gymnasium
stable-baselines3
